+++ integration-tests/src/deal-matcher.test.ts
@@ -13,33 +13,29 @@
 // TODO: more variates not only 1to1 match.
 import { assert, beforeAll, describe, expect, test } from "vitest";
 
-import {
-  type ContractsENV,
-  DealClient,
-  DealMatcherClient,
-  type GetMatchedOffersOut,
-} from "@fluencelabs/deal-ts-clients";
+import { type ContractsENV, DealClient } from "@fluencelabs/deal-ts-clients";
 import { ethers, JsonRpcProvider, type JsonRpcSigner, Wallet } from "ethers";
-import { DEFAULT_CONFIRMATIONS } from "./constants.js";
+import { CCStatus, DEFAULT_CONFIRMATIONS } from "./constants.js";
 import {
   createCommitments,
-  depositCapacity,
+  depositCollateral,
   registerMarketOffer,
 } from "./helpers.js";
+import { checkEvents } from "./confirmations.js";
+import { skipEpoch } from "./utils.js";
+import { config } from "dotenv";
+config({ path: [".env", ".env.local"] });
 
 // TODO: from env.
 
 const TEST_NETWORK: ContractsENV = "local";
-const TEST_RPC_URL = `http://local:8545`;
-const DEFAULT_SUBGRAPH_TIME_INDEXING = 300000;
+const TEST_RPC_URL = process.env.RPC_URL;
 const DEFAULT_TEST_TIMEOUT = 180000;
 // Test timeout should include:
 // - time for confirmations waits (1 confirmation is setup on anvil chain start)
 // - time for subgraph indexing
 // - time for core epoch
 // - other eps.
-// TODO: get core.epochDuration instead of 30000
-const TESTS_TIMEOUT = 120000 + 30000 + DEFAULT_SUBGRAPH_TIME_INDEXING;
 
 let provider: JsonRpcProvider;
 let signer: Wallet | JsonRpcSigner;
@@ -64,8 +60,6 @@ describe("#getMatchedOffersByDealId", () => {
   test(
     `Check that it matched successfully for 1:1 configuration.`,
     async () => {
-      const fromBlock = (await provider.getBlock("latest"))?.number;
-      assert(fromBlock, "from block isn't defined");
       const signerAddress = await signer.getAddress();
       const timestamp = (await provider.getBlock("latest"))?.timestamp;
       assert(timestamp, "Timestamp is defined");
@@ -75,13 +69,15 @@ describe("#getMatchedOffersByDealId", () => {
 
       console.info("---- Offer Creation ----");
       const marketContract = await contractsClient.getMarket();
+      const coreContract = await contractsClient.getCore();
       const capacityContract = await contractsClient.getCapacity();
+      const dealFactoryContract = await contractsClient.getDealFactory();
+      const epochDuration = await coreContract.epochDuration();
 
       const registeredOffer = await registerMarketOffer(
         marketContract,
         signerAddress,
         paymentTokenAddress,
-        fromBlock,
       );
 
       console.log("---- CC Creation ----");
@@ -89,24 +85,18 @@ describe("#getMatchedOffersByDealId", () => {
         capacityContract,
         signerAddress,
         registeredOffer.peers.map((p) => p.peerId),
-        fromBlock,
       );
 
       console.info("Deposit collateral for all sent CC...");
-      await depositCapacity(capacityContract, commitmentIds);
+      await depositCollateral(capacityContract, commitmentIds);
+      await skipEpoch(provider, epochDuration, 1);
+
+      for (const ccId of commitmentIds) {
+        const status = await capacityContract.getStatus(ccId);
+        assert(Number(status) === CCStatus.Active, "Status is not active");
+      }
 
       console.log("---- Deal Creation ----");
-      const marketAddress = await marketContract.getAddress();
-
-      const coreContract = await contractsClient.getCore();
-      const epochMilliseconds = (await coreContract.epochDuration()) * 1000n;
-      console.log(
-        `Wait for core epoch to pass: ${epochMilliseconds} milliseconds...`,
-      );
-      await new Promise((resolve) =>
-        setTimeout(resolve, Number(epochMilliseconds)),
-      );
-
       const minWorkersDeal = 1;
       const targetWorkersDeal = 1;
       const maxWorkerPerProviderDeal = 1;
@@ -126,21 +116,20 @@ describe("#getMatchedOffersByDealId", () => {
       expect(await paymentToken.balanceOf(signerAddress)).toBeGreaterThan(
         toApproveFromDeployer,
       );
-      expect(marketAddress).not.toBe("0x0000000000000000000000000000");
-      await paymentToken.approve(marketAddress, toApproveFromDeployer);
+      await paymentToken.approve(
+        await dealFactoryContract.getAddress(),
+        toApproveFromDeployer,
+      );
 
       console.info("Create deal that match default offer...");
-      const filter = marketContract.filters.DealCreated;
-      const lastDealsCreatedBefore = await marketContract.queryFilter(
-        filter,
-        fromBlock,
-      );
-      const deployDealTx = await marketContract.deployDeal(
+      const protocolVersion = 1;
+      const deployDealTx = await dealFactoryContract.deployDeal(
         {
           prefixes: "0x12345678",
           hash: ethers.encodeBytes32String(`appCID:${timestamp}`),
         },
         registeredOffer.paymentToken,
+        toApproveFromDeployer,
         minWorkersDeal,
         targetWorkersDeal,
         maxWorkerPerProviderDeal,
@@ -148,76 +137,47 @@ describe("#getMatchedOffersByDealId", () => {
         registeredOffer.effectors,
         0,
         [],
+        protocolVersion,
       );
       await deployDealTx.wait(DEFAULT_CONFIRMATIONS);
 
-      const lastDealsCreatedAfter = await marketContract.queryFilter(
-        filter,
-        fromBlock,
-      );
-      expect(lastDealsCreatedAfter.length - lastDealsCreatedBefore.length).toBe(
+      const [lastDealCreated] = await checkEvents(
+        dealFactoryContract,
+        dealFactoryContract.filters.DealCreated,
         1,
+        deployDealTx,
       );
-      const dealId =
-        lastDealsCreatedAfter[lastDealsCreatedAfter.length - 1]?.args.deal;
+
+      assert(lastDealCreated, "Deal is not created");
+
+      expect({
+        owner: lastDealCreated.args.owner,
+        minWorkers: lastDealCreated.args.minWorkers,
+        targetWorkers: lastDealCreated.args.targetWorkers,
+      }).toEqual({
+        owner: signerAddress,
+        minWorkers: BigInt(minWorkersDeal),
+        targetWorkers: BigInt(targetWorkersDeal),
+      });
+      const dealId = lastDealCreated?.args.deal;
       assert(dealId, "Deal ID is not defined");
 
-      console.info(`Wait till indexer ingesting blockchain state...`);
+      await skipEpoch(provider, epochDuration, 1);
 
-      const dealMatcherClient = new DealMatcherClient(
-        TEST_NETWORK,
-        `http://localhost:8000/subgraphs/name/fluence-deal-contracts`,
-      );
+      const peer = registeredOffer.peers[0];
+      assert(peer, "At least 1 peer should be defined");
+      const cuIds = [...(await marketContract.getComputeUnitIds(peer.peerId))];
 
-      let matchedOffersOut: GetMatchedOffersOut;
-
-      // eslint-disable-next-line @typescript-eslint/no-unnecessary-condition,no-constant-condition
-      while (true) {
-        try {
-          console.info(`Find matched offers for dealId: ${dealId}...`);
-          matchedOffersOut =
-            await dealMatcherClient.getMatchedOffersByDealId(dealId);
-          break;
-        } catch (e) {
-          console.log(e);
-          console.log("Waiting 5s for the next attempt");
-          await new Promise((resolve) => setTimeout(resolve, 5000));
-        }
-      }
-
-      console.log(matchedOffersOut);
-
-      // const dealId = "0xd79df1927718b3212fa6e126ec4ad2b3ee1263d9"
-      console.log("---- Deal Matching ----");
-
-      expect(matchedOffersOut.offers.length).toBe(1); // At least with one previously created offer it matched.
-
-      console.log(
-        "computeUnitsPerOffers",
-        matchedOffersOut.computeUnitsPerOffers,
-      );
-      const cuId = matchedOffersOut.computeUnitsPerOffers[0]?.[0];
-
-      // Additional check for status of matched CC from chain perspective
-      for (const commitmentId of commitmentIds) {
-        // e.g. 4 == Failed; 0 - Active.
-        expect(Number(await capacityContract.getStatus(commitmentId))).eq(0);
-      }
-
-      console.info(
-        `Match deal with offers structure proposed by indexer: ${JSON.stringify(matchedOffersOut)}...`,
-      );
+      console.info(`Match deal with offers structure`);
       const matchDealTx = await marketContract.matchDeal(
         dealId,
-        matchedOffersOut.offers,
-        matchedOffersOut.computeUnitsPerOffers,
+        [registeredOffer.offerId],
+        [cuIds],
       );
 
-      const matchDealTxReceipt = await matchDealTx.wait(DEFAULT_CONFIRMATIONS);
       await matchDealTx.wait(DEFAULT_CONFIRMATIONS);
-      console.log(matchDealTxReceipt);
       //   TODO: check further.
     },
-    TESTS_TIMEOUT,
+    DEFAULT_TEST_TIMEOUT,
   );
 });