+++ subgraph/src/mappings/market.ts
@@ -153,7 +153,6 @@ export function handleComputeUnitCreated(event: ComputeUnitCreated): void {
   offer.save();
 }
 
-// It updates Peer and Offer.
 export function handlePeerCreated(event: PeerCreated): void {
   const peer = new Peer(event.params.peerId.toHexString());
   const offer = Offer.load(event.params.offerId.toHexString()) as Offer;
@@ -169,7 +168,6 @@ export function handlePeerCreated(event: PeerCreated): void {
   // Init stats below.
   peer.computeUnitsTotal = 0;
   peer.computeUnitsInDeal = 0;
-  peer.computeUnitsInCapacityCommitment = 0;
   peer.save();
 }
 
@@ -278,7 +276,7 @@ export function handleComputeUnitRemovedFromDeal(
 
   // Upd stats.
   deal.matchedWorkersCurrentCount = deal.matchedWorkersCurrentCount - 1;
-  if (!computeUnit.workerId == null) {
+  if (computeUnit.workerId != null) {
     computeUnit.workerId = null;
     computeUnit.save();
     deal.registeredWorkersCurrentCount = deal.registeredWorkersCurrentCount - 1;