+++ test/deal/GetStatus.t.sol
@@ -21,7 +21,6 @@ contract GetStatus is Test {
     DeployDealSystem.Deployment deployment;
     TestCore testCore;
     TestDealContract dealContract;
-    Deal.DealStorage dealStorage;
 
     // ------------------ Test ------------------
     function setUp() public {
@@ -38,7 +37,11 @@ contract GetStatus is Test {
     function test_WhenWorkersCountIsNotMin() public {
         dealContract.setMinWorkers(10);
         dealContract.setWorkerCount(5);
-        dealContract.setMaxPaidEpoch(100);
+        dealContract.setPricePerWorkerEpoch(1 ether);
+        dealContract.setTargetWorkers(100);
+        dealContract.setMaxPaidEpoch(0);
+        testCore.setCurrentEpoch(101);
+        dealContract.setFreeBalance(dealContract.targetWorkers() * testCore.minDealDepositedEpoches() * dealContract.pricePerWorkerEpoch());
 
         assertEq(uint8(dealContract.getStatus()), uint8(IDeal.Status.NOT_ENOUGH_WORKERS), "Status mismatch");
     }
@@ -47,16 +50,35 @@ contract GetStatus is Test {
         dealContract.setMinWorkers(10);
         dealContract.setWorkerCount(10);
         testCore.setCurrentEpoch(101);
+        dealContract.setPricePerWorkerEpoch(1 ether);
         dealContract.setMaxPaidEpoch(100);
+        dealContract.setTargetWorkers(100);
+        dealContract.setFreeBalance(dealContract.targetWorkers() * testCore.minDealDepositedEpoches() * dealContract.pricePerWorkerEpoch());
 
         assertEq(uint8(dealContract.getStatus()), uint8(IDeal.Status.INSUFFICIENT_FUNDS), "Status mismatch");
+    }
+
+    function test_WhenSmallBalance() public {
+        dealContract.setMinWorkers(10);
+        dealContract.setWorkerCount(10);
+        dealContract.setPricePerWorkerEpoch(1 ether);
+        testCore.setCurrentEpoch(101);
+        dealContract.setMaxPaidEpoch(0);
+        dealContract.setTargetWorkers(100);
+
+        uint256 minBalance = dealContract.targetWorkers() * testCore.minDealDepositedEpoches() * dealContract.pricePerWorkerEpoch();
+        dealContract.setFreeBalance(minBalance - 1);
+
+        assertEq(uint8(dealContract.getStatus()), uint8(IDeal.Status.SMALL_BALANCE), "Status mismatch");
     }
 
     function test_WhenActive() public {
         dealContract.setMinWorkers(10);
         dealContract.setWorkerCount(10);
+        dealContract.setPricePerWorkerEpoch(1 ether);
         testCore.setCurrentEpoch(100);
         dealContract.setMaxPaidEpoch(100);
+        dealContract.setFreeBalance(dealContract.targetWorkers() * testCore.minDealDepositedEpoches() * dealContract.pricePerWorkerEpoch());
 
         assertEq(uint8(dealContract.getStatus()), uint8(IDeal.Status.ACTIVE), "Status mismatc");
     }
@@ -68,9 +90,14 @@ contract TestCore {
     function setCurrentEpoch(uint256 epoch) public {
         currentEpoch = epoch;
     }
+
+    function minDealDepositedEpoches() public pure returns (uint256) {
+        return 2;
+    }
 }
 
 contract TestDealContract is Deal {
+    uint256 private _freeBalance;
     //#region internal
     function _getConfigStorageTest() internal pure returns (ConfigStorage storage s) {
         bytes32 storageSlot = bytes32(uint256(keccak256("fluence.deal.storage.v1.config")) - 1);
@@ -100,12 +127,24 @@ contract TestDealContract is Deal {
         _getConfigStorageTest().globalCore = ICore(core);
     }
 
+    function getFreeBalance() public view override returns (uint256) {
+        return _freeBalance;
+    }
+
     function setWorkerCount(uint256 workerCount) public {
         _getWorkerManagerStorageTest().workerCount = workerCount;
     }
 
-    function setMinWorkers(uint256 minWorkers) public {
-        _getConfigStorageTest().minWorkers = minWorkers;
+    function setMinWorkers(uint256 minWorkers_) public {
+        _getConfigStorageTest().minWorkers = minWorkers_;
+    }
+
+    function setPricePerWorkerEpoch(uint256 price) public {
+        _getConfigStorageTest().pricePerWorkerEpoch = price;
+    }
+
+    function setTargetWorkers(uint256 targetWorkers_) public {
+        _getConfigStorageTest().targetWorkers = targetWorkers_;
     }
 
     function setMaxPaidEpoch(uint256 maxPaidEpoch) public {
@@ -114,5 +153,9 @@ contract TestDealContract is Deal {
 
     function setIsEnded(bool isEnded) public {
         _getDealStorageTest().isEnded = isEnded;
+    }
+
+    function setFreeBalance(uint256 freeBalance_) public {
+        _freeBalance = freeBalance_;
     }
 }