+++ ts-client/src/dealExplorerClient/dealExplorerClient.ts
@@ -636,7 +636,10 @@ export class DealExplorerClient {
         await this._dealRpcClient!.getFreeBalanceDealBatch([dealId])
       )[0];
       const effectors = serializeEffectors(deal.effectors);
-      const { whitelist, blacklist } = serializeDealProviderAccessLists(deal.providersAccessType, deal.providersAccessList);
+      const { whitelist, blacklist } = serializeDealProviderAccessLists(
+        deal.providersAccessType,
+        deal.providersAccessList,
+      );
       res = {
         ...serializeDealsShort(deal, { dealStatus, freeBalance }),
         pricePerWorkerEpoch: tokenValueToRounded(
@@ -873,6 +876,7 @@ export class DealExplorerClient {
       capacityCommitmentRpcDetails.unlockedRewards,
       capacityCommitmentRpcDetails.totalRewards,
       capacityCommitment.rewardWithdrawn,
+      capacityCommitment.delegator,
     );
   }
 
@@ -890,7 +894,9 @@ export class DealExplorerClient {
       providerId: peer.provider.id,
       offerId: peer.offer.id,
       computeUnitsInDeal: peer.computeUnitsInDeal,
-      computeUnitsInCapacityCommitment: peer.computeUnitsInCapacityCommitment,
+      computeUnitsInCapacityCommitment: peer.currentCapacityCommitment
+        ? peer.currentCapacityCommitment?.activeUnitCount
+        : 0,
       computeUnitsTotal: peer.computeUnitsTotal,
     };
   }
@@ -919,7 +925,7 @@ export class DealExplorerClient {
     if (
       data.peer == undefined ||
       data.peer.joinedDeals == undefined ||
-      data.peer.joinedDeals?.length == 0
+      data.peer.joinedDeals.length == 0
     ) {
       return {
         data: [],