+++ ts-client/src/dealExplorerClient/serializers/schemes.ts
@@ -108,18 +108,15 @@ export function serializeProviderShort(
   } as ProviderShort;
 }
 
-// It composes only compute units with linked workerIds.
 export function serializeComputeUnits(
   fetchedComputeUnits: Array<ComputeUnitBasicFragment>,
 ): Array<ComputeUnit> {
   const res: Array<ComputeUnit> = [];
   for (const fetched of fetchedComputeUnits) {
-    if (fetched.workerId) {
-      res.push({
-        id: fetched.id,
-        workerId: fetched.workerId,
-      });
-    }
+    res.push({
+      id: fetched.id,
+      workerId: fetched.workerId ?? undefined,
+    });
   }
   return res;
 }
@@ -234,6 +231,7 @@ export function serializeCapacityCommitmentDetail(
   unlockedRewards: bigint | null,
   totalRewards: bigint | null,
   rewardWithdrawn: bigint,
+  delegatorAddress: string,
 ): CapacityCommitmentDetail {
   const _totalRewards = totalRewards ? totalRewards : BigInt(0);
   const _unlockedRewards = unlockedRewards ? unlockedRewards : BigInt(0);
@@ -262,5 +260,9 @@ export function serializeCapacityCommitmentDetail(
       _totalRewards * BigInt(1 - rewardDelegatorRate),
     ),
     rewardsTotal: tokenValueToRounded(_totalRewards + rewardWithdrawn),
+    delegatorAddress:
+      delegatorAddress == "0x0000000000000000000000000000000000000000"
+        ? null
+        : delegatorAddress,
   };
 }